Rename folders for better logic
